Consolidates previous code
